Fixing issued, provided in code review
