+++ Balancyй/Balancyй/Models/Constants.swift
@@ -1,6 +1,23 @@
-import Foundation
+import UIKit
 
 struct Const {    
     static let tabBarTitles = ["","",""]
     static let tabBarImages = ["tasksTabBarIcon","nill","profileTabBarIcon"]
+}
+
+enum FontName {
+    static let light = "e-Ukraine-Light"
+    static let regular = "e-Ukraine-Regular"
+    static let medium = "e-Ukraine-Medium"
+}
+
+enum AppColor {
+    static let background = UIColor(red: 0.07, green: 0.07, blue: 0.07, alpha: 1)
+    static let button = UIColor(red: 0.85, green: 0.2, blue: 0.33, alpha: 1)
+    static let pageIndicatorTintColor = UIColor(red: 0.17, green: 0.17, blue: 0.18, alpha: 1)
+}
+
+enum ImageName {
+    static let logo = "logo"
+    static let addPhoto = "addPhotoImage"
 }
