--- Balancyй/Balancyй/User Settings/ImagePicking/Models/ImagePickingConstants.swift
@@ -1,25 +1,0 @@
-//
-//  ImagePickingConstants.swift
-//  Balancy
-//
-//  Created by  Toropov Oleksandr on 06.01.2024.
-//
-
-import Foundation
-import UIKit.UIColor
-
-enum FontName {
-    static let light = "e-Ukraine-Light"
-    static let regular = "e-Ukraine-Regular"
-    static let medium = "e-Ukraine-Medium"
-}
-
-enum AppColor {
-    static let background = UIColor(red: 0.07, green: 0.07, blue: 0.07, alpha: 1)
-    static let button = UIColor(red: 0.85, green: 0.2, blue: 0.33, alpha: 1)
-}
-
-enum ImageName {
-    static let logo = "logo"
-    static let addPhoto = "addPhotoImage"
-}
