+++ Balancyй/Balancyй/UserSettings/ImagePicking/Views/CategoryBalanceViewController.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import SwiftUI
 
 final class CategoryBalanceViewController: UIViewController {
